requirements.txt
Pillow >= 11.3.0
IPTCInfo3 >= 2.1.4